Create BaseEMSurvey and consolidate MT with ATEM
